requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pandas==2.0.3
scikit-learn==1.3.0
mlflow==2.9.1
boto3==1.28.5
python-dotenv==1.0.0
